backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "book-url-ingestion"
version = "0.1.0"
description = "Book URL Ingestion & Vector Indexing Pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "AI Book Team", email = "team@ai-book.example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "cohere>=5.0.0",
    "qdrant-client>=1.8.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0"
]

[project.urls]
Homepage = "https://github.com/ai-book/book-url-ingestion"
Repository = "https://github.com/ai-book/book-url-ingestion"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[project.scripts]
book-ingest = "main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]